api/requirements.txt
relations-pymysql==0.6.12
relations-restx==0.6.2
micro-logger==0.1.2
redis==4.4.0
prometheus-flask-exporter==0.18.6
ptvsd==4.3.2
coverage==5.2.1
pylint==2.5.3
